Change scrollDelta to int

+++ src/electron/electron/main/entities/UserInputEntity.ts
@@ -12,7 +12,7 @@ export class UserInputEntity extends BaseTrackedEntity {
   @Column('float', { nullable: false })
   movedDistance: number;
 
-  @Column('float', { nullable: false })
+  @Column('int', { nullable: false })
   scrollDelta: number;
 
   @Column('datetime', { nullable: false })
